Chapter 19 : ARTICLE: The Fundamental Equation of a Flying Saucer, With Applications
ABSTRACT
We derive a simple fundamental equation for a flying saucer. The
equation is:
v/G=k*M*f*a^2/r^2  (1)
where the v is the velocity of the craft, G the gravitational constant,
M the negative gravitational mass, a the diameter of the rotating ring,
f frequency of rotation, r the radius of the negative vector potential
bubble, and k a constant of value 1 with units of m/kg. From the
equation it is possible to deduce that:
1) inertia can be eliminated,
2) the minimal size of possible for a saucer, in agreement with eye
witnesses reports,
3) explain the EM effect of car headlights and engines dying near UFOs,
4) explain why UFOs do not cause sonic booms at supersonic speeds,
 5) and explain reports of levitation and loss of car control near a
UFO. Our only non-standard assumption is that the saucer posses negative
mass, an assumption which is NOT contrary to known physics principles.
Eliminating Inertia
Bondi (1) has show that the existence of negative gravitational mass
does not violate general relativity. This conclusion has not been
refuted.
Sciama in 1957 did the first calculation that showed that the total
matter in the universe caused inertia (2). His calculation was
non-relativistic. Einstein, although he never explicitly included Mach's
Principle as part General Relativity, says on page 102 of The Meaning of
Relativity that inertia was mediated by the time derivative of the
vector potential A, that is, by gravity (3). In a recent paper entitled
"On Mach's principle: Inertia as gravitation", a Spanish group did a
fully relativistic analysis with linearized Einstein equations. They
used the time derivative of the vector potential, h_(0*i,0), but also a
small scalar counter term h_(00*i), retarded potentials, and included
consideration of dark energy. They came within 10% of verifying Mach's
Principle.
Therefore the principle that gravitation causes inertia is well
established.
Following Sciama (4) and Jefimenko (5,6) we have the vector potential
A=G/c^2*intρ*v/r* d*V  (2)
where ρ is the mass density and v its velocity. To find the vector
potential created by the remote stars postulated by Mach's principle,
Sciama carries out the integration over the mass distribution of the
entire universe, assuming the universe to be receding with velocity -v.
The v being constant comes out of the integral
A=(v*G)/(c^2)(intρ/r* d*V)  (3)
with
Φ=G(intρ/r*  d*V)  (4)
We are left with
A=(v*Φ)/(c^2)  (5)
where Φ is the universal scalar gravitational potential. Sciama then
changes coordinates and has the test particle move with v while the
universe is stationary. The recent calculation (4) also uses this
technique. We are going to interpret this equation to mean that the
particle id subject to the A vector as it moves through the universes
scalar potential.
Positive and negative masses repel each other through their fields, and
the field has to be zero on the boundary. If a region of space existed
around a gravitationally negative mass, the scalar potential would be
negative and the A vector derived from this field would be anti-parallel
to the velocity
A=-|ϕ|*v/c^2  (6)
as opposed to parallel, A=Φ*v/c^2.
This is to be expected as Φ and A form a four-vector in relativity
A_μ=(A,ϕ)  (7)
so if ϕ is negative, so is A.
These parallel conditions are mutually exclusive; you cannot have A
parallel and ani-parallel to v at the same time. Therefore the positive
and negative gravity spaces would have to be disjoint.