A qqbar_t represents a real or complex algebraic number (an element of \(\overline{\mathbb{Q}}\)) by its unique reduced minimal polynomial in \(\mathbb{Z}[x]\) and an isolating complex interval. The precision of isolating intervals is maintained automatically to ensure that all operations on qqbar_t instances are exact.
This representation is useful for working with individual algebraic
numbers of moderate degree (up to 100, say). Arithmetic in this
representation is expensive: an arithmetic operation on numbers of
degrees m and n involves computing and then factoring an annihilating
polynomial of degree mn and potentially also performing numerical
root-finding. For doing repeated arithmetic, it is generally more
efficient to work with the ca_t type in a fixed number field. The
qqbar_t type is used internally by the ca_t type to represent the
embedding of number fields in \(\mathbb{R}\) or \(\mathbb{C}\) and to
decide predicates for algebraic numbers.

https://github.com/crowlogic/arb4j/issues/563